Remove link from confirmation and rename choose letter to singular
